5 46:33

+++ login/src/main/java/io/muzoo/ssc/webapp/servlet/DeleteUserServlet.java
@@ -9,6 +9,7 @@ import io.muzoo.ssc.webapp.Routable;
 import io.muzoo.ssc.webapp.model.User;
 import io.muzoo.ssc.webapp.service.SecurityService;
 import io.muzoo.ssc.webapp.service.UserService;
+import org.apache.commons.lang.StringUtils;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -47,15 +48,22 @@ public class DeleteUserServlet extends HttpServlet implements Routable {
                 User currentUser = userService.findByUsername(username);
 
                 User deletingUser = userService.findByUsername(request.getParameter("username"));
-                if (userService.deleteUserByUsername(deletingUser.getUsername())) {
+                if (StringUtils.equals(currentUser.getUsername(), deletingUser.getUsername())) {
                     // go to user list page
-                    request.getSession().setAttribute("hasError", false);
-                    request.getSession().setAttribute("message", "User is deleted");
-                } else {
-                    //err
                     request.getSession().setAttribute("hasError", true);
-                    request.getSession().setAttribute("message", "Unable to delete User");
+                    request.getSession().setAttribute("message", "You cannot delete own account");
+                } else {
+                    if (userService.deleteUserByUsername(deletingUser.getUsername())) {
+                        // go to user list page
+                        request.getSession().setAttribute("hasError", false);
+                        request.getSession().setAttribute("message", "User is deleted");
+                    } else {
+                        //err
+                        request.getSession().setAttribute("hasError", true);
+                        request.getSession().setAttribute("message", "Unable to delete User");
+                    }
                 }
+
             } catch (Exception e) {
                 //err
                 request.getSession().setAttribute("hasError", true);
